Add scripts to determine minimal project list for builds
Introduced a set of Bash scripts that help in determining the minimal list of Maven modules that need to be built. Also added an action to run these scripts. Additionally, the README file is updated to include usage details for the newly added action. The changes ensure only the projects affected by a given change-set are built, thus optimizing the build process.