Add trim function and utilize it in scripts
A new trim function has been added to remove leading and trailing spaces in strings. This function is now being used in several scripts to ensure that the processed data does not contain any unwanted or erroneous spaces. This makes sure that the final processed output is generated in a cleaner and more standardized format.