Reverse sort order in module listing
The order in which modules are listed has been reversed in the generate_dependency_map.sh script. Previously, the 'find' command was sorting the directories in ascending order, but now it sorts them in descending order. The adjustment was made by adding the -r option to the 'sort' command in the loop that processes each module.