Add recursion for dependent modules in project list script
The update to the create_project_list script now includes recursion for dependent modules. It also ensures that the affected module will always be included in the list to be built, regardless of if any dependents exist. Additionally, a new test case has been added to confirm the appropriate output when the changes reside in a transitive dependency.